valid sell (product|s) test

+++ src/main/java/com/indev/blackfriday/Capsule.java
@@ -1,7 +1,7 @@
 package com.indev.blackfriday;
 
 public class Capsule extends Product {
-    public Capsule(int quantity, int price ) {
+    public Capsule(int quantity, float price ) {
         super(quantity, price);
     }
 }
