+++ src/main/java/com/indev/blackfriday/Machine.java
@@ -1,7 +1,7 @@
 package com.indev.blackfriday;
 
 public class Machine extends Product {
-    public Machine(int quantity, int price) {
+    public Machine(int quantity, float price) {
         super(quantity, price);
     }
 }
